Search in the assembly where the abstract type is for concrete implementations.
Yes, dirty hack - but so was the previous implementation (i.e. the code would still have had to be in the same assembly).  This was caused by moving ShaderData out from under MATD.  It's only intended to work for "close-knit" situations, not more general abstract classes like ARCOLChunk, etc, so it's really not that bad.  Honest.